requirements.txt
langchain-community==0.0.29
langchain-core==0.1.33
langchain-text-splitters==0.0.1
chromadb==0.4.24
sentence-transformers==2.2.2
flask==3.0.2
langchain-groq==0.1.0
beautifulsoup4==4.12.2